Relaying the window's key events to the libgdx part
This is working but not clean and therefore unfinished.
Lots of console outs that were used to find out how to handle this are spamming the output right now.
Also, translating the AWT-keycode to the libgdx-keycode is done using a method that was blatantly stolen from the LwjglAWTInput.java source, which may or may not be a license issue.
Still, it works. Hence the commit. I'll clean up the code later.
Refs #8