antlr compilation and parser referencing work from gradle build script
Notice that I changed the target folder.

I'm not sure if that was really needed -- although I seemed to progress after I did that, the final changes to the build script might as well just have been completely unrelated to that. I don't see anything wrong with a common, shared source folder though, so I just leave it as it is.