Sprite color routine changed
The major issue when drawing elements so far was that blending was done using
the alpha channel and the sprite was simply colourized using the default
routines. This led to the issue of any displayed element being either dark or
with the specific colour at max (as colours were simply multiplied) and then
being blended with other colours below, which made it hard to actually specify
how an element was supposed to look like.

I modified the sprite shader to instead use a function that maps the sprites
colours like this:

 * black --> black

 * 50% grey --> specified colour

 * white --> white

Alpha works as before (original alpha * colour alpha).

This means that we can now specifiy colours for a sprite, have solid black
outlines present and also brighter, whiter parts inside there. This again
means that I do not have to rely on the alpha channel for brighter colours
(which was a hack anyway, as the whole thing would have turned out quite
differently for a black background). I therefore redid the sprites to match the
new colour scheme, which makes them seem more saturated now (which can,
however, be adjusted easily).

The result is that different field and droplet colours can be distinguished
much more easily, which was the goal of this branch in the first place.

Fixes #32