Adjacency now an overlay
The adjacency display now uses a second, overlaid sprite instead of the field's
color. I am *not* a fan of this solution as it essentially breaks the whole LOD
approach as the information can no longer be encoded in the 1-pixel-display of
the circuit.

This should be regarded more as a point for discussion than a preferred
solution. Because it isn't.