The Biochip's HashMap<Point,BiochipField> field is now private
With the Biochip no longer using an array, I will need to implement methods to retrieve
certain values (such as largest x coordinate etc.) in order to get some zooming
methods working again.

I used this as an excuse to also make the field private.

We will sooner or later need this in order to be able to catch and notify changes
in the Biochip and propagate them to the UI. The implementation of HashMap does
not make it easy to register event handlers (and it doesn't need to), so hiding
it within its parent class should be a sensible approach to make this easier in
the long run.

We should probably consider doing this for some other variables as well, but I guess
that'll have to wait.