Range check no longer fails for 0.
Previously, Range.inRange(int i) always returned false for i=0. As we are working
with a 0-based time index in the visualization, this seems kind of wrong. Changing
this only seems to fix #47, but I'm far from sure about any issues this may trigger
in other parts of the code.

Fixes #47