Frame rate is now limited to 60 fps max
When running the visualization on a higher spec GPU, the transitions become barely visible due to happening too fast.
This is a result of the transitions being linked to the framerate instead of time.

Now usually, the way to go would be to use the time instead of the frames to calculate the next state.
While this is *still* something we should consider doing, I also think that burning a CPU to 100% for offering 120+ frames per second does not make much sense for something that's supposed to be a tool.
I therefore added a framerate limiter that kicks in at 60+ fps and puts the thread to sleep before calculating the next frame.
The result is a consistent transition display and a much lower cpu footprint *as long as the computer is not at its limit*.