Tabs basically work
Now, this is still a rather messy state but the basic functionality is there.
Open file now opens a new tab with the new file.
Clicking tabs switches to the according file.

Things left to do:

 * There is currently a large empty space below the tabs due to me having to add a dummy panel to each tab which currently isn't size 0.
 * There is currently no way to close a tab
 * Opening a tab zooms the circuit back to the overview

Apart from these urgent things, tabs basically work now though.
This is totally not a side-by-side view of two circuits as, technically, there is still only one visualization that is simply told to display different circuits.