Disable automatically running checkstyle
The problem is that the IntelliJ IDE somehow switches the working directoy
so that checkstyle does not find the referenced suppressions.xml file.
Running checkstyle via the gradlew core/dektop:checkstyleMain tasks works
fine. Therefore, it has to be run manually from now on until I found a
way to make IntelliJ use the correct path.