No exception thrown to console upon closing last tab, fixes #80
The last tab being closed no longer throws an exception.
To bypass all those pesky null pointers, an empty DrawableCircuit is created and used as the current one instead of using null.
This may or may not be the best solution, but I don't have an issue with this and it seems to work just fine.