Coordinate display now auto-scales
The text for the coordinate system at the edge of the circuit now auto-scales with the zoom level.
For the record: I don't like this.
I think the text should remain fixed in size, at a pre-defined text size.
Still, as there's a ticket, this should serve as a proof-of-concept that it works... and sucks.
Please check this and see how it's not improving anything and simply results in extremely large text.

Refs #78