Desktop is nearly checkstyle compliant
The only two checkstyle errors remaining are two method length violations, which I'm not too sure we should fix.
Simply because the methods in question are

 * the method that sets up the left-hand panel with lots of very same-ish ui initialization methods and
 * the method that translates every single keycode from awt to libgdx codes.

I also modified the right curly brace rule because I prefer them to be on the same line on case of else and catch statements.