Fixes the problem of #95 that there was a NPE
For some reasons, the spawnSVGManager method wasn't called when
trying to save an SVG image. The method is necessary as some parts
of libgdx are not initialized when the object containing the manager
is created and, therefore, the manager can not be created.


Also adds a ton of debugging output that might (should?) be removed
later on.