Now displaying fluid names
This commit hopefully enables the display of fluid names when the according display option is set (which it currently is by default).
The main issue is that I have checked several examples which all did not seem to contain any fluid names, which is kind of bad for testing.
If null values are accepted, "null" is being printed and moved with the droplet.
So, while this *seems* to kind of work, I'd really need to test it.
@keszocze, which examples work for this case?

Refs #68