Clicking droplets hides them
Clicking a droplet now hides it from the circuit and places a droplet icon in the top-right corner of the visualization in order to be able to click it again to un-hide it.
This is imho a much better solution that a swing-ui element or something similar which would have resulted in quite some code overhad for less convenience.
Still, this is an *idea* how this could work and should be regarded as up for discussion.

Refs #68