Add 32 to the coordinateMultiplier.
The 32 contains the width of the cellBorder and the gap between the cells.Now the cells are placed correctly. Perhaps this value could be calculated, I will look into this later.
EDIT: Just saw that the svg source width is 256. So this should also be a reason to choose this value for the coordinateMultiplier.