Remove annoying drop from Dispenser svg and core files.
When exporting there was an annoying drop to the side of the dispenser.
I removed it from the svg file and commented it out in the core file. If
this is needed anywhere else please tell me. The application looks the
same, I think you use png s there and they don't have this drop.