Fixed coordinate offset when drawing HUD messages
This is all a little weird. It seems that the coordinate problem originated from
scaling the fonts. I can't really tell *why* scaling a font would *sometimes*
result in an offset (but mostly wouldn't).

In order to fix this, I attempted to calculate the font coordinates using the
more recent GlyphLayout structure. That, however, was only available in a more
recent version of libgdx, hence the update of the build script. Due to API
changes, I also had to change some other things throughout the project. While it
turned out that this didn't fix the issue at hand, I still left the whole
version switch as it is now, because there seemingly have been various changes
to the API and structure of libgdx and I'd figure we should do this switch
rather sooner than later, even if it's just for being on the same version as the
online API reference.

In order to still fix the coordinate offset, I disabled text scaling again. I
know that this seems to battle the symptom instead of the cause, but as I still
think scaling bitmap fonts is a terrible idea, it's the way to go imho.

Fixes #106