Files are no longer parsed if they are already loaded
While this may or may not be a good decision, the current behaviour seems odd.
So far, when loading a file (e.g. by clicking a tab), the file is always parsed but the parsing results were discarded when the file was loaded before.
Now, if previous circuits are to be taken, the file is no longer parsed at all.