New approach to net border drawing
I fiddled around with doing it in the shader but to no avail.
This is a new approach to keep the drawing of nets kind of inside the current framework without any overlaying structures.
The idea is that instead of e.g. drawing a new bounding box on top of the fields, the fields themselves draw the box by having their edges colourized in the net colour.
It's a little hackish but works.
Now, the current net example that I'm testing this with has exactly one net that stretches across the whole, square circuit.
This really isn't a proper testing ground.
So, @keszocze, could you try this again?

Refs #73