Net colours are no longer used 50% on bounding boxes
Beforehand, the nets' colours were blended with the original field colour.
While this meant that even for fields that were fully coloured in their nets' colours you could see other colour overlays, this also meant that they were kind of bland.
I don't consider this corner case to be that important though, and as it is now nets can be differentiated more easily.