Routes can now be drawn as one large arrow from the droplet to its target
This is all still very much WIP. Currently, the droplet only points toward its
target (so the source doesn't point toward the droplet). The sprite is just a
stretched version of the standard route arrow (which looks *very* off for longer
distances) and colours etc. all aren't implemented.

Yet, it's a working solution for large arrows, so a viable first step that
doesn't break anything. Just keep it in its own branch as there is no in-app way
to enable or disable this whole thing.

Refs #74