Interference region update
While this commit basically focuses on updating the interference display, it
also lays the foundation for a change I think we need to perform sooner or
later: the addition of a custom color class.

When I went into the whole field color calculation, I noticed that the colors
were off. The more colors were part of a calculation, the worse it was. I then
checked the libgdx reference and noticed that while they do offer a clamp
method, they clamp the color values after *all operations*, meaning that even
intermediate results cannot store any values > 1 or < 0. This is a huge issue
for a visualization project that basically consists of a bunch of colorized
sprites.

I therefore added a custom color class. So far, it doesn't do much apart from
offering the most basic functions, but I guess adding some stuff and
documentation shouldn't take too long. And ta-daa, color calculation works much
better. That doesn't mean that it necessarily looks better now, but at least
it's what you'd expect to get.

In other news, the Colors.INTERFERENCE_REGION_OVERLAP_COLOR is now only used if
a *droplet* enters another interference region, not if two regions overlap and
the overlapping interference region is now getting brighter with each region
that contributes to its resulting color.

Refs #119