Give priority to cell usage display
If the option for displaying the cell usage is set, a cell usage of > 0
will replace the text on top of the field.

Note that fields with zero usage are not altered in any way. This may lead
to the situation that there are different things displayed at the same time.
I chose this as I figured that people might still want to see the IDs of
dispensers.