Fixed dispenser drawing
@keszocze this should fix this whole dispenser drawing thing.
The issue was introduced in 4dee34078b4864e4f859c6db1dc3c951c308d847 when
you added the getMsgTexture() method to the DrawableField class.
The problem is that you refactored the draw() method to contain less code,
which was instead shifted to getMsgTexture(), including several instructions
that set the coordinates of the object.
This may have been considered a bad smell (as a get-method shouldn't alter
an internal state) but didn't matter as all draw() calls also called that
getter.
When you now overwrote the getter with a custom function for the
DrawableDispenser you didn't add a call to the super-method, thus resulting
in the coordinates of the field remaining unset.

Refs #151