DrawableLine.java is no longer a hellish mess
The DrawableLine class now uses the new BioVizSpriteBatch to be drawn. While it
does not yet utilize the depth sorting function, it is no longer relying on its
own drawing queue and is not used statically. Instead, it is now a simple
DrawableSprite. It still needs some cleanup, but it is already *much* better
than the weird generics-abusing mess it was before.