Make sure that cell usage is computed when needed.
Before this change, the cell usage was only computed when the option for
the colorfull displaying of the usage. If the actual numbers were requested,
no computation was performed. That meant that the option only worked after
the other option has been turned on at least once before.

Note that this commit does not contain the code for actually displaying the
numbers; just the computation is started.